Refactor. Add tests.
