+++ Core/Core/Planner/CalendarFilter/Model/Interactor/CalendarFilterCountLimit.swift
@@ -1,0 +1,53 @@
+//
+// This file is part of Canvas.
+// Copyright (C) 2024-present  Instructure, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+
+public enum CalendarFilterCountLimit {
+    case limited(Int)
+    case unlimited
+
+    var count: Int {
+        if case .limited(let count) = self {
+            return count
+        }
+        return .max
+    }
+}
+
+public extension Optional where Wrapped == AppEnvironment.App {
+
+    var isCalendarFilterLimitEnabled: Bool {
+        switch self {
+        case .teacher: return true
+        case .none, .parent, .student: return false
+        }
+    }
+}
+
+public extension Array where Element == CDEnvironmentSetting {
+
+    func calendarFilterCountLimit(
+        isCalendarFilterLimitEnabled: Bool
+    ) -> CalendarFilterCountLimit {
+        guard isCalendarFilterLimitEnabled else {
+            return .unlimited
+        }
+
+        let isExtendedFilterEnabled = isEnabled(.calendar_contexts_limit)
+        return .limited(isExtendedFilterEnabled ? 20 : 10)
+    }
+}
