+++ Core/Core/Planner/CalendarFilter/Model/Interactor/CalendarFilterInteractor.swift
@@ -1,0 +1,180 @@
+//
+// This file is part of Canvas.
+// Copyright (C) 2024-present  Instructure, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+
+import Combine
+import CombineSchedulers
+
+public protocol CalendarFilterInteractor: AnyObject {
+    var filters: CurrentValueSubject<[CDCalendarFilterEntry], Never> { get }
+    var filterCountLimit: CurrentValueSubject<CalendarFilterCountLimit, Never> { get }
+    var selectedContexts: CurrentValueSubject<Set<Context>, Never> { get }
+
+    func load(ignoreCache: Bool) -> AnyPublisher<Void, Error>
+    func updateFilteredContexts(_ contexts: [Context], isSelected: Bool) -> AnyPublisher<Void, Error>
+
+    func contextsForAPIFiltering() -> [Context]
+}
+
+public class CalendarFilterInteractorLive: CalendarFilterInteractor {
+    public let filters = CurrentValueSubject<[CDCalendarFilterEntry], Never>([])
+    public let filterCountLimit = CurrentValueSubject<CalendarFilterCountLimit, Never>(.unlimited)
+    public let selectedContexts = CurrentValueSubject<Set<Context>, Never>(Set())
+
+    private let observedUserId: String?
+    private let env: AppEnvironment
+    private let isCalendarFilterLimitEnabled: Bool
+    private let scheduler: AnySchedulerOf<DispatchQueue>
+    private let filterProvider: CalendarFilterEntryProvider
+    private var subscriptions = Set<AnyCancellable>()
+
+    required public init(
+        observedUserId: String?,
+        env: AppEnvironment = .shared,
+        filterProvider: CalendarFilterEntryProvider,
+        isCalendarFilterLimitEnabled: Bool = AppEnvironment.shared.app.isCalendarFilterLimitEnabled,
+        scheduler: AnySchedulerOf<DispatchQueue> = .main
+    ) {
+        self.observedUserId = observedUserId
+        self.env = env
+        self.filterProvider = filterProvider
+        self.isCalendarFilterLimitEnabled = isCalendarFilterLimitEnabled
+        self.scheduler = scheduler
+
+        observeUserDefaultChanges()
+    }
+
+    public func load(ignoreCache: Bool) -> AnyPublisher<Void, Error> {
+        guard let filterPublisher = filterProvider.make(ignoreCache: ignoreCache) else {
+            return Fail<Void, Error>(error: NSError.internalError())
+                .eraseToAnyPublisher()
+        }
+
+        return filterPublisher
+            .flatMap { [isCalendarFilterLimitEnabled] filters in
+                Self.fetchFilterLimit(
+                    isCalendarFilterLimitEnabled: isCalendarFilterLimitEnabled,
+                    ignoreCache: ignoreCache
+                )
+                .map { limit in
+                    (filters, limit)
+                }
+            }
+            .map { [sessionDefaults = env.userDefaults, observedUserId, isCalendarFilterLimitEnabled] (filters, limit) in
+                let selectedContexts: Set<Context>? = {
+                    if let loadedContexts = sessionDefaults?.calendarSelectedContexts(for: observedUserId) {
+                        return loadedContexts.removeUnavailableFilters(filters: filters)
+                    } else if isCalendarFilterLimitEnabled {
+                        return filters.defaultFilters(limit: limit)
+                    } else {
+                        return nil
+                    }
+                }()
+
+                return (filters, limit, selectedContexts)
+            }
+            .map { [weak self] (filters, limit, selectedContexts) in
+                guard let self else { return }
+                self.filters.send(filters)
+                filterCountLimit.send(limit)
+
+                if let selectedContexts {
+                    self.selectedContexts.send(selectedContexts)
+                    env.userDefaults?.setCalendarSelectedContexts(selectedContexts, observedStudentId: observedUserId)
+                }
+
+                return
+            }
+            .eraseToAnyPublisher()
+    }
+
+    public func updateFilteredContexts(_ contexts: [Context], isSelected: Bool) -> AnyPublisher<Void, Error> {
+        guard var defaults = env.userDefaults else {
+            return Fail(error: NSError.internalError()).eraseToAnyPublisher()
+        }
+
+        if isSelected,
+           case .limited(let limit) = filterCountLimit.value,
+           selectedContexts.value.count >= limit {
+            return Fail(error: NSError.internalError()).eraseToAnyPublisher()
+        }
+
+        return Future { [observedUserId, selectedContexts] promise in
+            var newSelectedContexts = defaults.calendarSelectedContexts(for: observedUserId) ?? Set()
+
+            if isSelected {
+                newSelectedContexts.formUnion(contexts)
+            } else {
+                newSelectedContexts.subtract(contexts)
+            }
+
+            defaults.setCalendarSelectedContexts(newSelectedContexts, observedStudentId: observedUserId)
+            selectedContexts.send(newSelectedContexts)
+            promise(.success(()))
+        }
+        .eraseToAnyPublisher()
+    }
+
+    public func contextsForAPIFiltering() -> [Context] {
+        Array(selectedContexts.value)
+    }
+
+    // MARK: - Private
+
+    private static func fetchFilterLimit(
+        isCalendarFilterLimitEnabled: Bool,
+        ignoreCache: Bool
+    ) -> AnyPublisher<CalendarFilterCountLimit, Error> {
+        let useCase = GetEnvironmentSettings()
+        return ReactiveStore(useCase: useCase)
+            .getEntities(ignoreCache: ignoreCache)
+            .map { $0.calendarFilterCountLimit(isCalendarFilterLimitEnabled: isCalendarFilterLimitEnabled) }
+            .eraseToAnyPublisher()
+    }
+
+    private func observeUserDefaultChanges() {
+        NotificationCenter
+            .default
+            .publisher(for: UserDefaults.didChangeNotification)
+            // We delay one cycle to avoid a crash occuring when the observed student changes
+            // which also modifies userdefaults violating "Exclusive Access to Memory".
+            .receive(on: scheduler)
+            .compactMap { [env, observedUserId] _ in
+                env.userDefaults?.calendarSelectedContexts(for: observedUserId)
+            }
+            .sink { [weak selectedContexts] newSelectedContexts in
+                selectedContexts?.send(newSelectedContexts)
+            }
+            .store(in: &subscriptions)
+    }
+}
+
+private extension Set where Element == Context {
+
+    func removeUnavailableFilters(filters: [CDCalendarFilterEntry]) -> Set<Element> {
+        let availableContexts = filters.map { $0.context }
+        return intersection(availableContexts)
+    }
+}
+
+private extension Array where Element == CDCalendarFilterEntry {
+
+    func defaultFilters(limit: CalendarFilterCountLimit) -> Set<Context> {
+        let contexts = sorted().map { $0.context }
+        return Set(contexts.prefix(limit.count))
+    }
+}
